Added new JS libs hook + separated sidebar
- Added a hook for including JS libraries, assigning it as an array to
Smarty's tpl. Also added a line of code into header.tpl for each library
to be displayed and included there. This line must be default in every
new upcoming template, because from now on ezRPG will be dependant on
jQuery lib as it comes to AJAX requests. THis is myt try to redo that
MSG string in URI, fetching it via AJAX instead.

- Separated left sidebar code into sidebar_left.tpl file.